requirements.txt
http://www.djangoproject.com/m/releases/1.5/Django-1.5c1.tar.gz
kombu==2.4.7
django-celery-email
django-fluent-dashboard
south
djangorestframework
django-private-files
IPy
django-singletons
django-extensions
django_transaction_signals
django-celery
markdown
-e git+https://github.com/alex/django-filter.git#egg=django-filter
https://bitbucket.org/glic3rinu/django-admin-tools-fixes/get/tip.tar.gz
https://bitbucket.org/glic3rinu/django-registration/get/tip.tar.gz
pygments
requests
-e git+https://github.com/madisona/django-google-maps.git#egg=django-google-maps
